Moved tests from old testfile to new test file
